+++ Decoder/c++/main.cpp
@@ -2,7 +2,7 @@
 #include "fst.hpp"
 
 int main(int argc, char const* argv[]) {
-    Fst fst(BeamSearch(stoi(argv[1]), stod(argv[2])), "../DecodingGraph-large.txt", "../labels.ciphones");
+    Fst fst(BeamSearch((uint)stoi(argv[1]), stod(argv[2])), "../DecodingGraph-large.txt", "../labels.ciphones");
 
     ifstream in;
     in.open("./activations_test/files.txt");
@@ -16,7 +16,7 @@ int main(int argc, char const* argv[]) {
         getline(activationsFile, shape);
         vector<string> dims;
         split(shape, dims);
-        vector<vector<double>> activations(stoi(dims[0]), vector<double>(stoi(dims[1])));
+        vector<vector<double>> activations((uint)stoi(dims[0]), vector<double>((uint)stoi(dims[1])));
         read2d(activationsFile, activations);
         vector<const Arc*> path = fst.decode(activations, 30.);
         for (const auto& arc : path) {
